reconx/requirements.txt
colorama==0.4.6
requests==2.31.0
dnspython==2.4.2
urllib3==2.0.7
python-nmap==0.7.1 